This PR contains a few small Montgomery backend adjustments:
- Usage of constant function `modulus_has_spare_bit` in `can_use_no_carry_mul_optimization` to avoid doubling the logic.
- Usage of `copy_from_slice` at the end of `mul_assign` operation in order to use memcpy for result assignment.
- Avoid `carry` mutable variable re-declaration during Montgomery reduction in `square_in_place` and reusing previous one for efficiency.