@Pratyush 2024-02-24
Thank you for the clean ups! If possible, could you provide benchmarks on how this affects the run time? I'm primarily worried about the `copy_from_slice` change.