@tcoratger 2024-02-24
@Pratyush Sure I've just run benchmark on BLS12-381 Fr Field 

## Multiplication

- With `copy_from_slice` (as proposed in this PR):
``` shell
Arithmetic for Bls12_381::Fr/Multiplication
                        time:   [26.303 ns 26.313 ns 26.325 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
```

- With `(a.0).0 = r` (before)
```shell
Arithmetic for Bls12_381::Fr/Multiplication
                        time:   [26.334 ns 26.366 ns 26.414 ns]
                        change: [+0.0261% +0.1255% +0.2409%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe
```
 
## Square

 - With `let mut carry = 0;` declaration (before)
 ```shell
Arithmetic for Bls12_381::Fr/Square
                        time:   [21.618 ns 21.642 ns 21.667 ns]
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
```

- Without carry redeclaration (as proposed in this PR):
```shell
Arithmetic for Bls12_381::Fr/Square
                        time:   [21.603 ns 21.619 ns 21.637 ns]
                        change: [-0.2895% -0.1441% -0.0177%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

In conclusion, I have the impression that we are winners on both aspects with this PR, even if it is a very, very small gain.